refactor: move util methods to util.ts
